requirements.txt
#tiktoken>=0.5.0
openai>=2.7.0
#pandas>=1.5.0
#matplotlib>=3.7.0
#plotly>=5.15.0
#scikit-learn>=1.2.0
#numpy>=1.21.0
python-dotenv>=1.1.0